Added in instructions and end game functions
Endgame and instructions are both functions that set a global boolean to determine what should be drawn in the draw function. When the user presses the instructions button, the game stops and the user goes to the instructions page. When the user presses the exit button, the game entirely stops.